[refactor] #102 본사 정산 facade 패키지 수정
